refactor formula helpers to chunks
